remove codec and add comments, make blob from EigendaBlobDAta public
